requirements.txt
Package    Version
---------- -------
awscrt     0.20.9
awsiotsdk  1.21.4
pip        23.0.1
setuptools 66.1.1
